[Refactor]Extract common code from `hadoop-ext` package
