+++ fe/fe-core/src/main/java/com/starrocks/connector/hadoop
@@ -1,0 +1,1 @@
+../../../../../../../../java-extensions/hadoop-ext/src/main/java/com/starrocks/connector/hadoop
