+++ fe/fe-core/src/main/java/com/starrocks/credential/CloudConfiguration.java
@@ -14,23 +14,16 @@
 
 package com.starrocks.credential;
 
-import com.google.common.base.Strings;
 import com.staros.proto.FileStoreInfo;
-import com.starrocks.StarRocksFE;
+import com.starrocks.connector.hadoop.HadoopExt;
 import com.starrocks.thrift.TCloudConfiguration;
 import com.starrocks.thrift.TCloudType;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.Path;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.HashMap;
 import java.util.Map;
-
-import static com.starrocks.credential.CloudConfigurationConstants.HDFS_CLOUD_CONFIGURATION_STRING;
-import static com.starrocks.credential.CloudConfigurationConstants.HDFS_CONFIG_RESOURCES;
-import static com.starrocks.credential.CloudConfigurationConstants.HDFS_CONFIG_RESOURCES_LOADED;
-import static com.starrocks.credential.CloudConfigurationConstants.HDFS_RUNTIME_JARS;
 
 public class CloudConfiguration {
     private static final Logger LOG = LogManager.getLogger(CloudConfiguration.class);
@@ -41,28 +34,15 @@ public class CloudConfiguration {
     public void toThrift(TCloudConfiguration tCloudConfiguration) {
         tCloudConfiguration.cloud_type = TCloudType.DEFAULT;
         Map<String, String> properties = new HashMap<>();
-        properties.put(HDFS_CONFIG_RESOURCES, configResources);
-        properties.put(HDFS_RUNTIME_JARS, runtimeJars);
-        properties.put(HDFS_CLOUD_CONFIGURATION_STRING, toConfString());
+        properties.put(HadoopExt.HDFS_CONFIG_RESOURCES, configResources);
+        properties.put(HadoopExt.HDFS_RUNTIME_JARS, runtimeJars);
+        properties.put(HadoopExt.HDFS_CLOUD_CONFIGURATION_STRING, toConfString());
         tCloudConfiguration.setCloud_properties_v2(properties);
     }
 
-    public static void addConfigResourcesToConfiguration(String configResources, Configuration conf) {
-        if (Strings.isNullOrEmpty(configResources)) {
-            return;
-        }
-        String[] parts = configResources.split(",");
-        for (String p : parts) {
-            Path path = new Path(StarRocksFE.STARROCKS_HOME_DIR + "/conf/", p);
-            LOG.debug(String.format("Add path '%s' to configuration", path.toString()));
-            conf.addResource(path);
-        }
-        conf.setBoolean(HDFS_CONFIG_RESOURCES_LOADED, true);
-    }
-
     public void applyToConfiguration(Configuration configuration) {
-        addConfigResourcesToConfiguration(configResources, configuration);
-        configuration.set(HDFS_CLOUD_CONFIGURATION_STRING, toConfString());
+        HadoopExt.addConfigResourcesToConfiguration(configResources, configuration);
+        configuration.set(HadoopExt.HDFS_CLOUD_CONFIGURATION_STRING, toConfString());
     }
 
     // Hadoop FileSystem has a cache itself, it used request uri as a cache key by default,
@@ -83,8 +63,8 @@ public class CloudConfiguration {
     }
 
     public void loadCommonFields(Map<String, String> properties) {
-        configResources = properties.getOrDefault(HDFS_CONFIG_RESOURCES, "");
-        runtimeJars = properties.getOrDefault(HDFS_RUNTIME_JARS, "");
+        configResources = properties.getOrDefault(HadoopExt.HDFS_CONFIG_RESOURCES, "");
+        runtimeJars = properties.getOrDefault(HadoopExt.HDFS_RUNTIME_JARS, "");
     }
 
     public String getCommonFieldsString() {
