+++ java-extensions/hadoop-ext/src/main/java/org/apache/hadoop/fs/FileSystem.java
@@ -34,6 +34,7 @@
  */
 package org.apache.hadoop.fs;
 
+import com.starrocks.connector.hadoop.HadoopExt;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.classification.InterfaceAudience;
